Splitting views file.
The views file has gotten very long and cumbersome. Splitting it into
separate files with dedicated purposes should make it eassier to manage.
