+++ routes18xxweb/routes18xxweb.py
@@ -1,7 +1,9 @@
-from flask import Flask
+from flask import Blueprint, Flask
 from flask_wtf.csrf import CSRFProtect
 
 app = Flask(__name__, template_folder='templates', static_folder='static')
 app.config.from_object("routes18xxweb.settings")
+
+game_app = Blueprint('game_app', __name__)
 
 csrf = CSRFProtect(app)
